Move orderedCloneDeep to InstanceDefinition.ts
InstanceDefinition is the only class that uses orderedCloneDeep. If
other classes need it later on, it can move to a shared location at that
time.
