+++ src/fhirtypes/common.ts
@@ -1,3 +1,4 @@
+import { isEmpty, cloneDeep } from 'lodash';
 import {
   StructureDefinition,
   PathPart,
@@ -31,10 +32,8 @@ import {
 } from '../fshtypes';
 import { FSHTank } from '../import';
 import { Type, Fishable } from '../utils/Fishable';
-import cloneDeep = require('lodash/cloneDeep');
 import { logger } from '../utils';
 import { FHIRId, idRegex } from './primitiveTypes';
-import { isEmpty, remove, pull } from 'lodash';
 
 export function splitOnPathPeriods(path: string): string[] {
   return path.split(/\.(?![^\[]*\])/g); // match a period that isn't within square brackets
@@ -594,46 +593,6 @@ export function getUrlFromFshDefinition(
     fhirType = 'StructureDefinition';
   }
   return `${canonical}/${fhirType}/${fshDefinition.id}`;
-}
-
-/**
- * Make a deep clone recursively, adding properties in the order expected for exported JSON.
- * If a list of keys is provided, use those properties from the input.
- * Otherwise, use all properties from the input.
- *
- * @param input - the value to clone
- * @param keys - optionally, the properties of the value to include in the clone
- * @returns {any} - a clone of the input, with reordered properties
- */
-export function orderedCloneDeep(input: any, keys: string[] = Object.keys(input)): any {
-  // non-objects should be cloned normally
-  // arrays should get a recursive call on their elements, but don't need reordering
-  if (typeof input !== 'object') {
-    return cloneDeep(input);
-  } else if (Array.isArray(input)) {
-    return input.map(element => orderedCloneDeep(element));
-  } else {
-    const underscoreKeys = remove(keys, key => key.startsWith('_'));
-    const orderedKeys: string[] = [];
-    const result: any = {};
-
-    keys.forEach(key => {
-      orderedKeys.push(key);
-      if (underscoreKeys.includes(`_${key}`)) {
-        orderedKeys.push(`_${key}`);
-        pull(underscoreKeys, `_${key}`);
-      }
-    });
-    underscoreKeys.forEach(key => {
-      orderedKeys.push(key);
-    });
-
-    orderedKeys.forEach(key => {
-      result[key] = orderedCloneDeep(input[key]);
-    });
-
-    return result;
-  }
 }
 
 const nameRegex = /^[A-Z]([A-Za-z0-9_]){0,254}$/;
